+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @under-control/forms
 
+## 1.6.17
+
+### Patch Changes
+
+- useForm is no more relaxed
+- Updated dependencies
+  - @under-control/inputs@1.6.17
+  - @under-control/core@1.6.17
+  - @under-control/validate@1.6.17
+
 ## 1.6.16
 
 ### Patch Changes