+++ packages/inputs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @under-control/inputs
 
+## 1.6.17
+
+### Patch Changes
+
+- useForm is no more relaxed
+- Updated dependencies
+  - @under-control/core@1.6.17
+
 ## 1.6.16
 
 ### Patch Changes